Made new login and register modals, could do with a few tweaks but they're pretty enough for now. Also stopped authors from liking own adventures

+++ functions.php
@@ -68,34 +68,48 @@ function addLike($postID)
     if ($loggedIn['user_group'] >= 1) {
         $oConn = LoginToDB();
 
-        $query = $oConn->prepare('SELECT * FROM Upvoted WHERE PostID = :postID AND Username = :username');
+        //Check if the user is the author of the article
+        $query = $oConn->prepare('SELECT Username FROM Adventures WHERE PostID = :postID AND Username = :username');
         $query->bindValue(':postID', $postID, PDO::PARAM_STR);
         $query->bindValue(':username', $username, PDO::PARAM_STR);
         $query->execute();
-        $rows = $query->fetchAll(PDO::FETCH_ASSOC); //grab all values that match
+        $results = $query->fetchAll(PDO::FETCH_ASSOC);
 
-        if ($rows) {
-            $update = $oConn->prepare('UPDATE Adventures SET Upvotes = Upvotes-1 WHERE PostID = :postID;DELETE FROM Upvoted WHERE PostID = :postID AND Username = :username');
-            $update->bindValue(':postID', $postID, PDO::PARAM_STR);
-            $update->bindValue(':username', $username, PDO::PARAM_STR);
+        if (!$results) {
+            //Check if the user has previously upvoted the post
+            $query = $oConn->prepare('SELECT * FROM Upvoted WHERE PostID = :postID AND Username = :username');
+            $query->bindValue(':postID', $postID, PDO::PARAM_STR);
+            $query->bindValue(':username', $username, PDO::PARAM_STR);
+            $query->execute();
+            $rows = $query->fetchAll(PDO::FETCH_ASSOC); //grab all values that match
 
-            if ($update->execute()) {
-                return json_encode(array('success' => 'Upvote removed'));
+            //If has previous upvote, remove upvote, else add upvote
+            if ($rows) {
+                $update = $oConn->prepare('UPDATE Adventures SET Upvotes = Upvotes-1 WHERE PostID = :postID;DELETE FROM Upvoted WHERE PostID = :postID AND Username = :username');
+                $update->bindValue(':postID', $postID, PDO::PARAM_STR);
+                $update->bindValue(':username', $username, PDO::PARAM_STR);
+
+                if ($update->execute()) {
+                    return json_encode(array('success' => 'Upvote removed'));
+                } else {
+                    return json_encode(array('error' => 'Something went wrong'));
+
+                }
             } else {
-                return json_encode(array('error' => 'Something went wrong'));
+                $update = $oConn->prepare('UPDATE Adventures SET Upvotes = Upvotes+1 WHERE PostID = :postID;INSERT INTO Upvoted VALUES(NULL, :postID, :username)');
+                $update->bindValue(':postID', $postID, PDO::PARAM_STR);
+                $update->bindValue(':username', $username, PDO::PARAM_STR);
 
+                if ($update->execute()) {
+                    return json_encode(array('success' => 'Upvote added'));
+                } else {
+                    return json_encode(array('error' => 'Something went wrong'));
+
+                }
             }
         } else {
-            $update = $oConn->prepare('UPDATE Adventures SET Upvotes = Upvotes+1 WHERE PostID = :postID;INSERT INTO Upvoted VALUES(NULL, :postID, :username)');
-            $update->bindValue(':postID', $postID, PDO::PARAM_STR);
-            $update->bindValue(':username', $username, PDO::PARAM_STR);
+            return json_encode(array('error' => 'Cannot like own adventure'));
 
-            if ($update->execute()) {
-                return json_encode(array('success' => 'Upvote added'));
-            } else {
-                return json_encode(array('error' => 'Something went wrong'));
-
-            }
         }
 
     } else {
